+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add impersonation flag category and better NIP-56 mapping.
 - Added a filter button to the Home tab that lets you browse all notes on a specific relay.
 - Add a Tap to Refresh button in empty profiles.
 - Update the reply count shown below each Note in a Feed.